Refactor type casting in various source files
The type casting in SCBE_Emit.cpp, SCBE_X64.cpp, SymTableHash.cpp, and ByteCode.cpp source files have been refactored to improve consistency and readability. This includes the correction of static type casting operations, like uint32_t for numBits and name lengths, instead of simple subtraction and comparison operations.

+++ src/SymTableHash.cpp
@@ -57,7 +57,7 @@ void SymTableHash::addElem(SymbolName* data, uint32_t crc)
 
     buffer[idx].hash       = crc;
     buffer[idx].symbolName = data;
-    maxLength              = max(maxLength, (uint32_t) data->name.length());
+    maxLength              = max(maxLength, data->name.length());
     count += 1;
 }
